+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/java/JavaRestCall.java
@@ -162,9 +162,8 @@ public class JavaRestCall extends AbstractRestCall
         {
             builder.setHeader("Content-Type", contentType);
         }
-
+        
         return builder.build();
-
     }
 
     protected BodyPublisher createBodyPublisher()